+++ src/app/types/index.d.ts
@@ -130,6 +130,8 @@ export interface IAppConfig {
     headingLineSize: number; // mode for display of heading line -1 = default
     iconScale: number; // scale to apply to self Vessel icon
     rangeCircles: boolean; //display range circles
+    rangeCircleCount: number; // number of circles to display
+    rangeCircleMinZoom: number; // min zoom level where they are displayed
     aisStaleAge: number; // time since last update in ms (6 min)
     aisMaxAge: number; // time since last update in ms (9 min)
     aisWindApparent: boolean;